README: Fix installation instructions to use sudo
The install command before the commit works fine in a Dockerfile but
when running it interactively most users are in a user account instead
of a root account. In addition, they want to see command output to check
for errors instead of curl silently failing as it happens when running
the command without root permissions.